windows: fix size of `ssize_t`
Previously, `long long` was used unconditionally to define `ssize_t` on
Windows. This meant it was always a 64-bit value. However, typically
the size type is a 32-bit value on 32-bit machines and 64-bit value on
64-bit machines.

This didn't cause any issues until it was used on 32-bit Windows build
of `libvalkey-py`. CPython defines the `ssize_t` type as well as
`libvalkey`, but it defines the size to be 32 bit on 32 bit machines
which caused a compilation error.

This commit changes the typedef to be `SSIZE_T` on Windows. Hence, it
will be aligned with everything else.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>