include: make LIBVALKEY_SONAME a string
Previously it was a literal `1.2.1-dev` that had to be put under
`clang-format off` section because clang-format treated that expression
as subtraction and added spaces around minus sign.

Making it a string, however, required the users of `LIBVALKEY_SONAME`,
namely `Makefile` and `CMakeLists.txt`, to ignore the possible quotes in
the variable.

As a benefit, `LIBVALKEY_SONAME` can be used in the code as well in case
such thing is needed.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>