Update and unify version tags on installed shared libraries (#196)
Previously a CMake-build and a pure Makefile-build gave different
results when installing the built shared libraries.
This PR unifies the result, i.e. a CMake build and the provided Makefile
gives the same results.

The PR also changes the version tags on installed libraries to include
the patch version, and
the `SONAME` in the built libraries are changed to only include the
major version as described in some
[docs](https://tldp.org/HOWTO/Program-Library-HOWTO/shared-libraries.html).

The naming of the version defines in `valkey.h` are improved, e.g.
`LIBVALKEY_MAJOR` is changed to `LIBVALKEY_VERSION_MAJOR`, and the
SONAME define is removed (MAJOR used instead..)

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>