Fix the long-blocking read for Valkey RDMA.
During the valkey benchmark, when we set the data size as 16KB, the
benchmark will be blocking on GET. The reason behind is that RDMA
event is edge triggered and every incoming data has to be read totally
instead of read partially. We modify the valkeyBufferRead to enable a
total read over the RDMA buffer when the conneciton type is RDMA. To
realize this logic we need to keep read the buffer until the read
result data size is smaller than the attempt read size.

In addition, we need to deal with a corner case for the logic above.
If the message happened to be 16KB and the first read finish all the
available data, the second read will be triggered and block there until
return an error. To solve this problem, we make read non-blocking
except for the first read.

Signed-off-by: Ruihong Wang <ruihong@google.com>