Fix the long-blocking read for Valkey RDMA
During the benchmark, when we set the data size as 16KB or more, the
benchmark will be blocked on GET. The reason behind is that RDMA
event is edge triggered and every incoming data has to be read totally
instead of read partially. To solve this problem, we add 'read_zc' in
valkey.h to enable read with zero copy. This function enables the
valkeyBufferRead to feed all available data into c->reader from RDMA
buffer at every coming RDMA event.

Signed-off-by: wang4996 <wang4996@purdue.edu>