Add option to select a logical database (#244)
Set the option `select_db` in `valkeyClusterOptions` to select a logical
database after a successful connect.
The option is supported in both the Synchronous and Asynchronous API,
and the `SELECT` command is only sent when configured to a non-zero
value.

Additionally adds new flags to the clusterclients `[--select-db NUM]`
and update tests.

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>