Avoid heap buffer overflow in valkeyAsyncFormattedCommand (#245)
`valkeyAsyncFormattedCommand` now returns `VALKEY_ERR` instead of
crashing when the command length, or content, is faulty.

Adds validation of the parsed length to make sure we don't read past the
buffer end. The internal `nextArgument` function now takes a new function
argument, the buffer length, to be able to do the validation.

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>
Co-authored-by: michael-grunder <michael.grunder@gmail.com>