Fix the long-blocking read for Valkey RDMA. (#233)
During the valkey benchmark, when we set the data size as 16KB, the
benchmark will be blocking on GET. The reason behind is that RDMA event
is edge-triggered and hence every incoming data has to be read totally
instead of read partially. Otherwise, the benchmark will be blocked at
the event loop . We modify the valkeyBufferRead to enable a total read
over the RDMA buffer when the conneciton type is RDMA. To realize this
logic we need to keep read the buffer until the read result data size is
smaller than the attempt read size.

In addition, we need to deal with a corner case for the logic above. If
the message happened to be 16KB and the first read finishes all the
available data, the second read will be triggered and block there until
return an error. To solve this problem, we make read non-blocking except
for the first read.

An alternative approach to solve this problem could be making the
message sender chunk the message into 16KB blocks (equal to the stack
buffer) and generate the PollIn signal every 16KB data arrival. This
approach seems to also improve the performance of RDMA commands with
very large data size. According to our benchmark result. this pipelined
RDMA transfer can improve "SET" command throughput by 1.4X. I am okay
with both solutions.

Signed-off-by: wang4996 <wang4996@purdue.edu>