Support slotmap updates using CLUSTER NODES in RESP3 (#262)
When the option `VALKEY_OPT_USE_CLUSTER_NODES` is enabled the slot map
is updated using the `CLUSTER NODES` command. If a user has enabled
RESP3 by sending the `HELLO 3` command the `CLUSTER NODES` response is a
verbatim string instead of a bulk string.
Update so we handle this reply type as well.

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>